fix: Merge conflicts and binance requestor chain with gdax

+++ app/src/main/java/urbanutility/design/kaleidoscope/datatypes/OrderType.java
@@ -1,8 +1,5 @@
 package urbanutility.design.kaleidoscope.datatypes;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /**
  * Created by Jason on 1/13/2018.
  */
@@ -12,10 +9,10 @@ public class OrderType {
     public String exchange; /* Binance, Kraken... */
     public String symbol;
     public String type; /* Buy or Sell */
-    public long price;
-    public long btcUsdRate;
-    public long amount;
-    public long txFee;
+    public double price;
+    public double btcUsdRate;
+    public double amount;
+    public double txFee;
     public String time; /* ISO 8601 format: "yyyy-MM-dd'T'HH:mm:ssZ" */
 
     public OrderType () {
